back/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
supabase
python-dateutil==2.9.0.post0
google-genai
openai==1.35.13
httpx==0.27.0
